Map args to INPUT vars
This closes #6.

The entrypoint.sh allows us to map args back to INPUT_* type variables.
This, in turn, permits the usage of the core libraries to get inputs.

Eventually, we may want to bring in a CLI framework, but at the
time of this writing, this isn't necessary.

It may be the case once we start publishing the actions to docker
repositories.